Move tests to *_test as siblings of implementations
