+++ src/actor/Actor_test.ts
@@ -12,7 +12,7 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import { Actor, hookup, lookup } from "../../actor/Actor.js";
+import { Actor, hookup, lookup } from "./Actor.js";
 
 declare var window: { Mocha: Mocha.MochaGlobals; assert: Chai.Assert };
